+++ circumflex-core/src/main/scala/router.scala
@@ -131,12 +131,12 @@ class RequestRouter {
   /**
    * Sends a file.
    */
-  def sendFile(file: File) = FileResponse(file)
+  def sendFile(file: File): FileResponse = FileResponse(file)
 
   /**
    * Sends a file with Content-Disposition: attachment and specified UTF-8 filename.
    */
-  def sendFile(file: File, filename: String) = {
+  def sendFile(file: File, filename: String): FileResponse = {
     attachment(filename)
     sendFile(file)
   }